Try to make archive retrival work (#32)
* add docs how you can use your local storage to work

* add secrets support

* add secrets support and update packages

* add new endpoint so an azure event grid can send if a blob was changed in the tier

* implement the frontend side to retrive blobs

* cleanup new file

* cleanup function

* let retrival of blobs rock!

* try to fix the pipeline. (this solution would be sad)

* cleanup devocontainers file and update it to the latest and greatest

* add github actions extensions

* try to use build artifacts

* start to make changes to main.bicep

* try to make the yaml more robust

* fix main.yaml

* try to fix the deployment scripts

* fix dev container

* try to maybe improve deployment for linux

* try to fix deploy.ps1 paths for zip deploy

* now the bicep file is able to deploy for the current PR the app

* fix current deployment

* make it possible to deploy from main path too

* add deploy.ps1 to the github release

* make the bicep stuff more production ready

* change hostname to base url

* deploy events for PR branches automatically

* remove a hack we needed befor

* try to fix deployment

* try to fix supported version numbers

* try to fix the infra deployment

* fix bicep template for the event subscriptions

* fix pr commenting

* try to fix the comments

* make our deployment more modular

* update the build pipline to be more efficent

* fix build pipeline

* fix dotnet build / publish.

* try to fix azure deployment

* fix deploy web app script

* try to fix deployment

* fix even more parameters

* try to use the correct URL to download things while PR is running

* try it again

* add debugging

* debug more

* maybe this helps

* fix deployment

* try it again

* fix web app deployment

* try to add more debug information

* fix a long implemented bug

* fix event deployment

* try it again

* fix deployment

* fix cors

* try to fix the output of the correct URL

* try to fix the output

* replace static strings with variables

* fix output for main branch too

* fix main deployment

* add https to the comment url

* make the change back to use variables

* fix some of the small bugs at the download page

* improve the extension collection

* try to clear the list better

* make improvements to the development container so it's easier to try out things localy

* fix a list bug on the download page.
Improve the logging a bit

+++ build/deploy.ps1
@@ -24,6 +24,9 @@ param (
     [Parameter(Mandatory = $false)]
     [string]
     $workingDir = "$pwd/temp/temp/",
+    [Parameter(Mandatory = $false)]
+    [switch]
+    $cleanupWorkingDir,
     # Set a branch name that is used in the deciption of the envrionment
     [Parameter(Mandatory = $false)]
     [string]
@@ -104,7 +107,7 @@ if((Get-Item $apiBuildOutput).PSIsContainer){
     $apiZipToUpload = "$workingDir/api.zip";
     Compress-Archive "$apiBuildOutput\*" -DestinationPath $apiZipToUpload -Force;
 }else{
-    $apiZipToUpload = $apiBuildOutput;
+    $apiZipToUpload = "$workingDir/$apiBuildOutput";
 }
 $appZipToUpload = ([string]::Empty);
 if((Get-Item $appBuildOutput).PSIsContainer){
@@ -112,7 +115,7 @@ if((Get-Item $appBuildOutput).PSIsContainer){
     $appZipToUpload = "$workingDir/app.zip";
     Compress-Archive "$appBuildOutput\*" -DestinationPath $appZipToUpload -Force;
 }else{
-    $appZipToUpload = $appBuildOutput;
+    $appZipToUpload = "$workingDir/$appBuildOutput";
 }
 $apiHash = (Get-FileHash $apiZipToUpload -Algorithm MD5).Hash;
 
@@ -216,11 +219,13 @@ while (
     $response;
     Start-Sleep 2;
 }
-if((Test-Path $workingDir)){
+if($cleanupWorkingDir -and (Test-Path $workingDir)){
     Remove-Item -Recurse -Force $workingDir;
 }
 if($response.response.deploymentStatus -ne 'Succeeded'){
     throw "The deployment failed. The reason was: $($response.response.errorDetails)";
 }
 $response.response.siteUrl;
+# for github actions
 Write-Output "::set-output name=SiteUrl::$($response.response.siteUrl)";
+
